+++ packages/maestro_cli/lib/src/features/doctor/doctor_command.dart
@@ -2,9 +2,9 @@ import 'package:args/command_runner.dart';
 
 import '../../common/common.dart';
 
-class ConfigCommand extends Command<int> {
+class DoctorCommand extends Command<int> {
   @override
-  String get name => 'config';
+  String get name => 'doctor';
 
   @override
   String get description => 'Show configuration.';